third_party/WebKit/LayoutTests/fast/dom/Window/window-methods-construct.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
  test(t => {
    assert_throws(null, function() { new window.blur(); });
  }, "Construct call to window methods throws.");
</script>
